Started a PeerConnection class which allows a better remote video management
It is now possible to share the screen either replacing the local video
stream or creating a new stream, which will open a separate video on the
remote peer.
To select either one, set the global variable screenReplaceVideo, accordingly.
Sharing a video, which also opens up a separate video on the remote peer
can be stopped by a new button.