Improvements to chat message and use of IndexedDB
Chat messages are now saved to and retrieved from IndexedDB based
on the token of the call. Subsequent calls with the same token
loads the existing messages.
Allow several lines in the message formatting line by line to fit
on the canvas rectangle
Login and remote party name saved and retrieved for the message
Date is formatted according to how old it is
Input replaced by text area for input automatic formatting and
scrolling
Canvas, text area and button placed on a grid to allow better
positioning and overlay
css from the chat in a separate file: chat.css